requirements.txt
# Core
apache-airflow==3.1.3
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-amazon==8.14.0

# Streaming
kafka-python==2.0.2
confluent-kafka==2.3.0

# Processing
pyspark==3.5.0
pandas==2.1.4
numpy==1.26.3
pyarrow==14.0.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Dashboard
streamlit==1.31.0
plotly==5.18.0
altair==5.2.0

# Data Quality
great-expectations==0.18.8

# Utils
python-dotenv==1.0.0
boto3==1.34.14        # For S3 (Data Lake)
requests==2.31.0
faker==22.2.0         # Sample data generation
loguru==0.7.2         # Logging

# Testing
pytest==7.4.4
pytest-mock==3.12.0
